START: State Machine / closer
