+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/internal/Controller.java
@@ -5,8 +5,6 @@ import io.confluent.parallelconsumer.PollContextInternal;
 import io.confluent.parallelconsumer.state.WorkManager;
 import lombok.Getter;
 import lombok.NonNull;
-import lombok.Setter;
-import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.consumer.ConsumerRebalanceListener;
 
@@ -23,12 +21,9 @@ import java.util.function.Consumer;
 import java.util.function.Function;
 
 import static io.confluent.csid.utils.BackportUtils.isEmpty;
-import static io.confluent.csid.utils.BackportUtils.toSeconds;
-import static io.confluent.csid.utils.StringUtils.msg;
 import static io.confluent.parallelconsumer.internal.AbstractParallelEoSStreamProcessor.addInstanceMDC;
 import static io.confluent.parallelconsumer.internal.State.*;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
-import static java.util.concurrent.TimeUnit.SECONDS;
 import static lombok.AccessLevel.*;
 
 /**
@@ -41,14 +36,6 @@ public class Controller<K, V> {
 
     @Getter(PROTECTED)
     protected final ParallelConsumerOptions<K, V> options;
-
-    /**
-     * The run state of the controller.
-     *
-     * @see State
-     */
-    @Setter
-    private State state = State.unused;
 
     private Optional<Future<Boolean>> controlThreadFuture = Optional.empty();
 
@@ -473,119 +460,6 @@ public class Controller<K, V> {
     }
 
     /**
-     * Close the system, without draining.
-     *
-     * @see State#draining
-     */
-    @Override
-    public void close() {
-        // use a longer timeout, to cover for evey other step using the default
-        Duration timeout = DrainingCloseable.DEFAULT_TIMEOUT.multipliedBy(2);
-        closeDontDrainFirst(timeout);
-    }
-
-    @Override
-    @SneakyThrows
-    public void close(Duration timeout, DrainingCloseable.DrainingMode drainMode) {
-        if (state == closed) {
-            log.info("Already closed, checking end state..");
-        } else {
-            log.info("Signaling to close...");
-
-            switch (drainMode) {
-                case DRAIN -> {
-                    log.info("Will wait for all in flight to complete before");
-                    transitionToDraining();
-                }
-                case DONT_DRAIN -> {
-                    log.info("Not waiting for remaining queued to complete, will finish in flight, then close...");
-                    transitionToClosing();
-                }
-            }
-
-            waitForClose(timeout);
-        }
-
-        if (controlThreadFuture.isPresent()) {
-            log.debug("Checking for control thread exception...");
-            Future<?> future = controlThreadFuture.get();
-            future.get(timeout.toMillis(), MILLISECONDS); // throws exception if supervisor saw one
-        }
-
-        log.info("Close complete.");
-    }
-
-    private void waitForClose(Duration timeout) throws TimeoutException, ExecutionException {
-        log.info("Waiting on closed state...");
-        while (!state.equals(closed)) {
-            try {
-                Future<Boolean> booleanFuture = this.controlThreadFuture.get();
-                log.debug("Blocking on control future");
-                boolean signaled = booleanFuture.get(toSeconds(timeout), SECONDS);
-                if (!signaled)
-                    throw new TimeoutException("Timeout waiting for system to close (" + timeout + ")");
-            } catch (InterruptedException e) {
-                // ignore
-                log.trace("Interrupted", e);
-            } catch (ExecutionException | TimeoutException e) {
-                log.error("Execution or timeout exception while waiting for the control thread to close cleanly " +
-                        "(state was {}). Try increasing your time-out to allow the system to drain, or close without " +
-                        "draining.", state, e);
-                throw e;
-            }
-            log.trace("Still waiting for system to close...");
-        }
-    }
-
-    private void doClose(Duration timeout) throws TimeoutException, ExecutionException, InterruptedException {
-        log.debug("Starting close process (state: {})...", state);
-
-        log.debug("Shutting down execution pool...");
-        List<Runnable> unfinished = workerThreadPool.shutdownNow();
-        if (!unfinished.isEmpty()) {
-            log.warn("Threads not done count: {}", unfinished.size());
-        }
-
-        log.debug("Awaiting worker pool termination...");
-        boolean interrupted = true;
-        while (interrupted) {
-            log.debug("Still interrupted");
-            try {
-                boolean terminationFinishedWithoutTimeout = workerThreadPool.awaitTermination(toSeconds(timeout), SECONDS);
-                interrupted = false;
-                if (!terminationFinishedWithoutTimeout) {
-                    log.warn("Thread execution pool termination await timeout ({})! Were any processing jobs dead locked (test latch locks?) or otherwise stuck?", timeout);
-                }
-            } catch (InterruptedException e) {
-                log.error("InterruptedException", e);
-                interrupted = true;
-            }
-        }
-        log.debug("Worker pool terminated.");
-
-        // last check to see if after worker pool closed, has any new work arrived?
-        workMailbox.processWorkCompleteMailBox(Duration.ZERO);
-
-        //
-        commitOffsetsThatAreReady();
-
-        // only close consumer once producer has committed it's offsets (tx'l)
-        log.debug("Closing and waiting for broker poll system...");
-        brokerPollSubsystem.closeAndWait();
-
-        maybeCloseConsumer();
-
-        producerManager.ifPresent(x -> x.close(timeout));
-
-        log.debug("Close complete.");
-        this.state = closed;
-
-        if (this.getFailureCause() != null) {
-            log.error("PC closed due to error: {}", getFailureCause(), null);
-        }
-    }
-
-    /**
      * Early notify of work arrived.
      * <p>
      * Only wake up the thread if it's sleeping while polling the mail box.
@@ -657,82 +531,6 @@ public class Controller<K, V> {
         }
     }
 
-    /**
-     * To keep things simple, make sure the correct thread which can make a commit, is the one to close the consumer.
-     * This way, if partitions are revoked, the commit can be made inline.
-     */
-    private void maybeCloseConsumer() {
-        if (isResponsibleForCommits()) {
-            consumer.close();
-        }
-    }
-
-    private boolean isResponsibleForCommits() {
-        return (committer instanceof ProducerManager);
-    }
-
-    private void transitionToDraining() {
-        log.debug("Transitioning to draining...");
-        this.state = State.draining;
-        notifySomethingToDo();
-    }
 
 
-    private void drain() {
-        log.debug("Signaling to drain...");
-        brokerPollSubsystem.drain();
-        if (!isRecordsAwaitingProcessing()) {
-            transitionToClosing();
-        } else {
-            log.debug("Records still waiting processing, won't transition to closing.");
-        }
-    }
-
-    private void transitionToClosing() {
-        log.debug("Transitioning to closing...");
-        if (state == State.unused) {
-            state = closed;
-        } else {
-            state = State.closing;
-        }
-        notifySomethingToDo();
-    }
-
-    @Override
-    public void pauseIfRunning() {
-        if (this.state == State.running) {
-            log.info("Transitioning parallel consumer to state paused.");
-            this.state = State.paused;
-        } else {
-            log.debug("Skipping transition of parallel consumer to state paused. Current state is {}.", this.state);
-        }
-    }
-
-    @Override
-    public void resumeIfPaused() {
-        if (this.state == State.paused) {
-            log.info("Transitioning parallel consumer to state running.");
-            this.state = State.running;
-            notifySomethingToDo();
-        } else {
-            log.debug("Skipping transition of parallel consumer to state running. Current state is {}.", this.state);
-        }
-    }
-
-    public boolean isClosedOrFailed() {
-        boolean closed = state == State.closed;
-        boolean doneOrCancelled = false;
-        if (this.controlThreadFuture.isPresent()) {
-            Future<Boolean> threadFuture = controlThreadFuture.get();
-            doneOrCancelled = threadFuture.isDone() || threadFuture.isCancelled();
-        }
-        return closed || doneOrCancelled;
-    }
-
-    /**
-     * @return if the system failed, returns the recorded reason.
-     */
-    public Exception getFailureCause() {
-        return this.failureReason;
-    }
 }
